+++ src/Randomizer/LongPartyPatch.js
@@ -1,0 +1,8 @@
+import { utils } from 'randomtools-js';
+
+const LongPartyPatch = {
+    name: "Long Party",
+    entries: utils.standardPatchLoader(["4E83: 14"]),
+};
+
+export default LongPartyPatch;